Added Service and Dao

+++ src/main/java/com/projectStore/projectStore/entity/Description.java
@@ -4,9 +4,7 @@ import lombok.Data;
 
 import javax.persistence.*;
 
-/**
- * Created by 111 on 25.03.2019.
- */
+
 @Data
 @Entity
 @Table(name = "description")
@@ -15,7 +13,7 @@ public class  Description {
     @Column(name = "description_id")
     @SequenceGenerator(name = "description_seq", sequenceName = "description_seq", allocationSize = 1)
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "description_seq")
-    private int descriptionId;
+    private Long descriptionId;
 
     @Column(name = "description_goods")
     private String descriptionGoods;
