+++ src/main/java/com/projectStore/projectStore/entity/GoodsType.java
@@ -4,9 +4,7 @@ import lombok.Data;
 
 import javax.persistence.*;
 
-/**
- * Created by 111 on 25.03.2019.
- */
+
 @Data
 @Entity
 @Table(name = "goods_type")
@@ -15,7 +13,7 @@ public class GoodsType {
     @Column(name = "goods_type_id")
     @SequenceGenerator(name = "goods_type_seq", sequenceName = "goods_type_seq", allocationSize = 1)
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "goods_type_seq")
-    private int goodsTypeId;
+    private Long goodsTypeId;
 
     @Column(name = "type_name")
     private String typeName;
